+++ src/IFECStrategy.hpp
@@ -18,7 +18,7 @@ public:
     virtual std::vector<std::vector<uint8_t>> decode(
         const std::map<uint16_t, std::vector<uint8_t>>& received_symbols, 
         size_t K_data_symbols,
-        uint32_t block_id) = 0;
+        uint16_t block_id) = 0;
 };
 
 } // namespace udpworm
