+++ src/LDPCFEC.cpp
@@ -88,7 +88,7 @@ std::vector<std::vector<uint8_t>> LDPCFEC::encode(
 std::vector<std::vector<uint8_t>> LDPCFEC::decode(
     const std::map<uint16_t, std::vector<uint8_t>>& received_symbols,
     size_t K_data_symbols,
-    uint32_t block_id) {
+    uint16_t block_id) {
     (void)block_id;
     if (received_symbols.empty() || K_data_symbols == 0) {
         return {};
