+++ src/LDPCFEC.hpp
@@ -16,7 +16,7 @@ public:
     std::vector<std::vector<uint8_t>> decode(
         const std::map<uint16_t, std::vector<uint8_t>>& received_symbols,
         size_t K_data_symbols,
-        uint32_t block_id) override;
+        uint16_t block_id) override;
 
 private:
     size_t data_symbols_;
